--- web/src/api/party.functions.ts
@@ -1,1 +1,0 @@
-export const requestSomeStuff = () => {};
