--- web/src/components/admin/user-management/user-list/UserList.css
@@ -1,5 +1,0 @@
-@media screen and (max-width: 900px) {
-  .user-list-mobile-hidden {
-    display: none;
-  }
-}
