+++ web/src/weekly-view/footer/Footer.css
