+++ web/src/weekly-view/footer/Footer.tsx
