+++ web/src/weekly-view/party-card/PartyCard.css
