+++ web/src/weekly-view/party-card/PartyCard.tsx
